Crowbar added
Added the crowbar on the devmap.
Added bg picture with crowbar in it.
Added the ability to find crowbar in room #5.
Room #5 shows a crowbar if the player doesn't have one already. No crowbar appears if the player does have one already.
Can use the crowbar to open the cracked wall.
Opened wall gives different text depending on if tnt or crowbar is used to open it.